+++ penman/transform.py
@@ -6,11 +6,11 @@ Tree and graph transformations.
 from typing import Union, Dict, Set, List, Tuple
 import logging
 
-from penman.types import Variable, Target, BasicTriple
+from penman.types import (Variable, Target, BasicTriple, Node)
 from penman.exceptions import ModelError
 from penman.epigraph import (Epidatum, Epidata)
 from penman.surface import (Alignment, RoleAlignment, alignments)
-from penman.tree import (Tree, Node, is_atomic)
+from penman.tree import (Tree, is_atomic)
 from penman.graph import (Graph, CONCEPT_ROLE)
 from penman.model import Model
 from penman.layout import (Push, POP, appears_inverted, get_pushed_variable)